join_service: add check on ak received from attester
When an attester try to join the framework, the join_service check
if the AK correspondant is already registered in its database and
if it is valid.
If the ak is not present it store it in its database.
Before this operation MUST be added the check on the EK received
together with the AK, which is not present yet.