join_service: add tpm_makecredential for AK verification
The join service performs tpm_makecredential using the public part of
the EK, a secret and the AK name. This value will be used again on the
attester which has to prove that the AK sent is actually stored in the
TPM associated with the EK provided. This proof is the fact that the
attester can find the secret.

Only an RSA key can be used as currently implemented, but trivial
modification can be made in order to support ECC keys too.