attester: add tpm_activatecredential
Added the tpm_activate credential on the attester. This operation
permits to assert the fact that the provided AK is stored on the TPM
associated with the provided EK.

The steps implemented are the following:

1. $ tpm2_startauthsession --policy-session -S session.ctx

2. $ tpm2_policysecret -S session.ctx -c e

3. $ tpm2_activatecredential -Q -c ak.ctx -C <EK_handle> -i mkcred.out \
-o actcred.out -P"session:<session_ctx_path>"

4. $ tpm2_flushcontext <session_ctx_path>

The complete process is implemented reading the mkcred_output, produced
by the join_service, from a memory buffer and processing it throught the
activate_credential. The activate_credential then produces as output the
secret chosen by the join_service and store it in another memory buffer
in order to send it back to the join_service.